# Conflicts:
#	composer.json
#	composer.lock
#	src/SbWereWolf/XmlNavigator/Convertation/FastXmlToArray.php
#	src/SbWereWolf/XmlNavigator/Convertation/IFastXmlToArray.php
#	src/SbWereWolf/XmlNavigator/Convertation/XmlConverter.php
#	src/SbWereWolf/XmlNavigator/Extraction/ElementExtractor.php
#	src/SbWereWolf/XmlNavigator/Navigation/XmlElement.php
#	test/Integration/DebugTest.php

+++ src/SbWereWolf/XmlNavigator/Extraction/ElementComposer.php
@@ -10,7 +10,7 @@ namespace SbWereWolf\XmlNavigator\Extraction;
 class ElementComposer
 {
     /**
-     * @param array $elems
+     * @param array<int,array<string,array<string,int>>> $elems
      * @return int
      */
     protected static function extractBaseDepth(array $elems): int
@@ -22,7 +22,7 @@ class ElementComposer
         }
         $base = 0;
         if ($first) {
-            $base = current($first)[ElementExtractor::DEPTH];
+            $base = (int)current($first)[ElementExtractor::DEPTH];
         }
 
         return $base;
